14020231-1312

+++ Migrations/ApplicationDbContextModelSnapshot.cs
@@ -108,11 +108,11 @@ namespace WoundClinic_WPF.Migrations
 
             modelBuilder.Entity("WoundClinic_WPF.Models.Dressing", b =>
                 {
-                    b.Property<byte>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("tinyint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<byte>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<string>("DressingName")
                         .IsRequired()
@@ -143,8 +143,8 @@ namespace WoundClinic_WPF.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<byte>("DressingId")
-                        .HasColumnType("tinyint");
+                    b.Property<int>("DressingId")
+                        .HasColumnType("int");
 
                     b.Property<int>("Price")
                         .HasColumnType("int");
